requirements.txt
attrs==25.3.0
aws-cdk-lib==2.190.0
aws-cdk.asset-awscli-v1==2.2.232
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==41.2.0
aws-lambda-powertools==1.31.1
aws-solutions-constructs.aws-apigateway-lambda==2.84.0
aws-solutions-constructs.core==2.84.0
aws-xray-sdk==2.14.0
black==22.12.0
boto3==1.20.32
boto3-stubs==1.20.32
botocore==1.23.54
botocore-stubs==1.37.37
cattrs==24.1.3
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
constructs==10.4.2
coverage==7.8.0
distlib==0.3.9
fastjsonschema==2.21.1
filelock==3.18.0
flake8==4.0.1
flake8-black==0.3.6
flake8-builtins==1.5.3
flake8-comprehensions==3.16.0
flake8-debugger==4.1.2
flake8-fixme==1.1.1
flake8-isort==4.2.0
flake8-variables-names==0.0.5
identify==2.6.9
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
isort==5.13.2
jmespath==0.10.0
jsii==1.111.0
mccabe==0.6.1
mypy==0.961
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pre-commit==2.21.0
publication==0.0.3
py==1.11.0
pycodestyle==2.8.0
pydantic==1.10.21
pyflakes==2.4.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-aws4auth==1.3.1
s3transfer==0.5.2
six==1.17.0
toml==0.10.2
typeguard==2.13.3
types-awscrt==0.26.1
types-setuptools==57.4.18
types-toml==0.10.8.20240310
typing_extensions==4.13.2
urllib3==1.26.20
virtualenv==20.30.0
wrapt==1.17.2
